# Fusion group optimizer