--- a/paddle/fluid/framework/ir/fusion_group/code_generator.cc
+++ b/paddle/fluid/framework/ir/fusion_group/code_generator.cc
@@ -108,7 +108,7 @@ std::vector<OperationExpression> CodeGenerator::ConvertToExpressions(
       std::vector<int> output_ids;
       std::vector<std::string> output_names =
           OperationMap::Instance().Get(op->Type()).output_names;
-      bool intermediate_state = false;
+      std::unordered_map<int, bool> intermediate_state;
 
       for (auto& name : output_names) {
         PADDLE_ENFORCE_NE(
@@ -116,12 +116,14 @@ std::vector<OperationExpression> CodeGenerator::ConvertToExpressions(
             platform::errors::InvalidArgument(
                 "Output(%s) of operation %s is not set.", name, op->Type()));
         output_ids.push_back(var_ids[op->Output(name)[0]]);
+        bool enable_intermediate = false;
         for (auto* n : intermediate_out_nodes) {
           if (n->Name() == op->Output(name)[0]) {
-            intermediate_state = true;
+            enable_intermediate = true;
             break;
           }
         }
+        intermediate_state[var_ids[op->Output(name)[0]]] = enable_intermediate;
       }
 
       std::string lhs_type = ExtractDataType(node->outputs);
@@ -206,7 +208,8 @@ std::set<int> CodeGenerator::DistilIntermediateIds(
   // Use std::set to remove the reptead id and get a ordered list.
   for (size_t i = 0; i < expressions.size(); i++) {
     for (auto id : expressions[i].GetOutputIds()) {
-      if (expressions[i].GetIntermediateState()) intermediate_ids.insert(id);
+      auto intermediate_state = expressions[i].GetIntermediateState();
+      if (intermediate_state[id]) intermediate_ids.insert(id);
     }
   }
   return intermediate_ids;