--- a/paddle/fluid/framework/ir/fusion_group/code_generator.h
+++ b/paddle/fluid/framework/ir/fusion_group/code_generator.h
@@ -51,6 +51,7 @@ class CodeGenerator {
   // we get the parameter list code for the expression information
   std::string EmitParameters(
       const std::set<int>& input_ids, const std::set<int>& output_ids,
+      const std::set<int>& intermediate_ids,
       const std::unordered_map<int, std::string>& dtypes) const;
 
   std::string EmitComputeBody(