--- a/paddle/fluid/framework/ir/fusion_group/code_generator_helper.h
+++ b/paddle/fluid/framework/ir/fusion_group/code_generator_helper.h
@@ -47,7 +47,7 @@ class OperationExpression {
   explicit OperationExpression(std::string op_type, std::vector<int> input_ids,
                                std::vector<int> output_ids,
                                std::string rhs_type, std::string lhs_type,
-                               bool intermediate_state)
+                               std::unordered_map<int, bool> intermediate_state)
       : op_type_(op_type),
         input_ids_(input_ids),
         output_ids_(output_ids),
@@ -56,7 +56,9 @@ class OperationExpression {
         intermediate_state_(intermediate_state) {}
 
   std::string GetOpType() const { return op_type_; }
-  bool GetIntermediateState() const { return intermediate_state_; }
+  std::unordered_map<int, bool> GetIntermediateState() const {
+    return intermediate_state_;
+  }
   std::vector<int> GetInputIds() const { return input_ids_; }
   std::vector<int> GetOutputIds() const { return output_ids_; }
   std::string GetRHSType() const { return rhs_type_; }
@@ -81,7 +83,7 @@ class OperationExpression {
   AttributeMap attr_;
   std::string rhs_type_;
   std::string lhs_type_;
-  bool intermediate_state_;
+  std::unordered_map<int, bool> intermediate_state_;
 };
 
 class TemplateVariable {