--- a/paddle/fluid/framework/ir/fusion_group/subgraph.h
+++ b/paddle/fluid/framework/ir/fusion_group/subgraph.h
@@ -19,6 +19,8 @@ limitations under the License. */
 #include <unordered_set>
 #include <vector>
 #include "paddle/fluid/framework/ir/fusion_group/operation.h"
+#include "paddle/fluid/framework/ir/graph.h"
+#include "paddle/fluid/framework/ir/graph_traits.h"
 #include "paddle/fluid/framework/ir/node.h"
 #include "paddle/fluid/framework/ir/subgraph_detector.h"
 
@@ -64,6 +66,7 @@ class SubGraph {
   }
 
   int GetType() const { return type_; }
+  bool RemoveIntermediateOut() { return !save_intermediate_out_; }
 
   void SetFuncName(std::string func_name) { func_name_ = func_name; }
   std::string GetFuncName() const { return func_name_; }
@@ -133,30 +136,40 @@ class SubGraph {
         }
       }
     }
+    return output_vars_all;
+  }
 
-    if (save_intermediate_out_) {
-      return output_vars_all;
-    }
+  std::vector<Node*> GetIntermediateOutVarNodes() {
+    return intermediate_out_nodes_;
+  }
+  void DetectIntermediateOutWithGraph(Graph* graph) {
+    auto graph_nodes = graph->Nodes();
 
-    std::vector<Node*> output_vars_outside;
-    for (auto* n : output_vars_all) {
-      // If one of the var_node's outputs is the input of some operator
-      // outside the subgraph, it is considered the output var node of the
-      // subgraph.
-      bool is_found = true;
-      if (n->outputs.size() == 0U) {
-        is_found = false;
-      }
-      for (auto* out : n->outputs) {
-        if (!Has(out)) {
-          is_found = false;
+    for (auto* n : SortedNodes()) {
+      bool enable_remove = true;
+
+      if (n && n->IsVar() && n->Var()) {
+        for (auto* node : graph_nodes) {
+          if (node->IsOp() && !Has(node)) {
+            auto inputs = node->inputs;
+            for (auto* in : inputs) {
+              if (in == n) {
+                enable_remove = false;
+              }
+            }
+          }
+          if (!enable_remove) {
+            break;
+          }
         }
+      } else {
+        enable_remove = false;
       }
-      if (!is_found) {
-        output_vars_outside.push_back(n);
+
+      if (enable_remove) {
+        intermediate_out_nodes_.push_back(n);
       }
     }
-    return output_vars_outside;
   }
 
  private:
@@ -218,6 +231,7 @@ class SubGraph {
   bool save_intermediate_out_{true};
 
   std::unordered_set<Node*> nodes_set_;
+  std::vector<Node*> intermediate_out_nodes_{};
   bool is_sorted_{false};
   std::vector<Node*> sorted_nodes_;
 };